End gen 4 moves

+++ include/menu_indicators.h
@@ -29,14 +29,14 @@ struct CursorStruct
     u8 palNum;
 };
 
-u8 AddScrollIndicatorArrowPairParameterized(u32 arrowType, s32 commonPos, s32 firstPos, s32 secondPos, s32 fullyDownThreshold, s32 tileTag, s32 palTag, u16 *scrollOffset);
-u8 AddScrollIndicatorArrowPair(const struct ScrollArrowsTemplate * template, u16 * a1);
-void RemoveScrollIndicatorArrowPair(u8);
-u8 ListMenuAddCursorObjectInternal(const struct CursorStruct *cursor, u32 cursorKind);
-u8 ListMenuAddRedOutlineCursorObject(const struct CursorStruct *cursor);
-void ListMenuUpdateRedOutlineCursorObject(u8 taskId, u16 x, u16 y);
-void ListMenuRemoveRedOutlineCursorObject(u8 taskId);
-void ListMenuRemoveCursorObject(u8 taskId, u32 cursorKind);
-void ListMenuUpdateCursorObject(u8 taskId, u16 x, u16 y, u32 cursorKind);
+u32 AddScrollIndicatorArrowPairParameterized(u32 arrowType, s32 commonPos, s32 firstPos, s32 secondPos, s32 fullyDownThreshold, s32 tileTag, s32 palTag, u16 *scrollOffset);
+u32 AddScrollIndicatorArrowPair(const struct ScrollArrowsTemplate * template, u16 * a1);
+void RemoveScrollIndicatorArrowPair(u32);
+u32 ListMenuAddCursorObjectInternal(const struct CursorStruct *cursor, u32 cursorKind);
+u32 ListMenuAddRedOutlineCursorObject(const struct CursorStruct *cursor);
+void ListMenuUpdateRedOutlineCursorObject(u32 taskId, u32 x, u32 y);
+void ListMenuRemoveRedOutlineCursorObject(u32 taskId);
+void ListMenuRemoveCursorObject(u32 taskId, u32 cursorKind);
+void ListMenuUpdateCursorObject(u32 taskId, u32 x, u32 y, u32 cursorKind);
 
 #endif //GUARD_MENU_INDICATORS_H
