+++ include/pokedex_area_markers.h
@@ -10,8 +10,8 @@ struct PAM_TaskData
 };
 
 void SetAreaSubsprite(s32 i, s32 whichArea, struct Subsprite * subsprites);
-void Dtor_PokedexAreaMarkers(u8 taskId);
-u8 Ctor_PokedexAreaMarkers(u16 species, u16 tilesTag, u8 palIdx, u8 y);
-u8 PokedexAreaMarkers_Any(u8 taskId);
+void Dtor_PokedexAreaMarkers(u32 taskId);
+u32 Ctor_PokedexAreaMarkers(u32 species, u32 tilesTag, u32 palIdx, u32 y);
+u32 PokedexAreaMarkers_Any(u32 taskId);
 
 #endif //GUARD_POKEDEX_AREA_MARKERS_H
